backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
python-dotenv==1.0.1
dj-database-url==2.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
drf-yasg==1.21.7
django-storages==1.14.2
boto3==1.34.34
celery==5.3.6
redis==5.0.1
stripe==7.13.0